bunch of updates
bunch of updates to directory, and changed about me content but someone will need to add the logos and give it a lookover for spelling/grammar. added a contact link to the menu and added it to the bottom of the map in place of the "meals must be eaten on site" since that is now stated in the about.